Rename non-public files for easier readability
Renamed non-public files from signed_video to sv_ for improved readability
