+++ lib/src/sv_defines.h
