+++ lib/src/sv_h26x_internal.h
@@ -23,8 +23,8 @@
 
 #include <stdbool.h>  // bool
 
-#include "signed_video_defines.h"  // svrc_t
-#include "signed_video_internal.h"  // gop_info_t, gop_state_t, MAX_HASH_SIZE
+#include "sv_defines.h"  // svrc_t
+#include "sv_internal.h"  // gop_info_t, gop_state_t, MAX_HASH_SIZE
 
 #define METADATA_TYPE_USER_PRIVATE 25
