+++ tests/check/test_helpers.c
@@ -34,9 +34,9 @@
 #include "lib/src/includes/signed_video_common.h"
 #include "lib/src/includes/signed_video_openssl.h"
 #include "lib/src/includes/signed_video_sign.h"
-#include "lib/src/signed_video_h26x_internal.h"  // parse_nalu_info(), kUuidSignedVideo
-#include "lib/src/signed_video_internal.h"  // _signed_video_t, UUID_LEN, ATTR_UNUSED
-#include "lib/src/signed_video_tlv.h"  // tlv_find_tag()
+#include "lib/src/sv_h26x_internal.h"  // parse_nalu_info(), kUuidSignedVideo
+#include "lib/src/sv_internal.h"  // _signed_video_t, UUID_LEN, ATTR_UNUSED
+#include "lib/src/sv_tlv.h"  // tlv_find_tag()
 
 #define RSA_PRIVATE_KEY_ALLOC_BYTES 2000
 #define ECDSA_PRIVATE_KEY_ALLOC_BYTES 1000
